Refactored ScanFiles out to Directory Service.
